Finished Loop Detection and ability to remove Edges
I finalized the loop detection method to include detecting if the loop had a way to check if all of the Special Nodes(Straight, SharpTurn, and WideTurn) were also contained in the loop.

Previously there was no way to remove edges but now if there is currently an edge between two nodes and the player moves between those nodes the edge will be removed.